requirements.txt
easydict==1.9
imageio==2.9.0
imutils==0.5.4
keras==2.6.0
numpy==1.20.0
opencv-contrib-python==3.4.8.29
opencv-python==3.4.9.31
Pillow==7.0.0
scikit-image==0.18.2
scikit-learn==0.22.2
scipy==1.4.1
tensorflow-gpu==2.6.0
tensorflow-estimator==2.6.0
tqdm==4.62.0